JieGou_Admin/pyproject.toml
[project]
name = "jiegou-admin"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "booktype",
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "hiredis>=3.2.1",
    "mysqlclient>=2.2.7",
    "pyjwt>=2.10.1",
]
